+++ packages/evaluate/src/weathergen/evaluate/io/wegen_reader.py
@@ -30,12 +30,12 @@ from weathergen.common.config import (
     load_run_config,
 )
 from weathergen.common.io import zarrio_reader
-from weathergen.evaluate.io.data._get_data_orchestration import (
+from weathergen.evaluate.io.data.get_data_orchestration import (
     _apply_postprocessing,
     _dispatch_reads,
     _reassemble_fsteps,
 )
-from weathergen.evaluate.io.data._raw_io_orchestration import (
+from weathergen.evaluate.io.data.raw_io_orchestration import (
     _build_raw_io_state,
     get_data_raw_impl,
     get_data_raw_zip_impl,
